+++ lib/render_kit/rendering_helper_monkey_patch.rb
@@ -4,9 +4,9 @@ require "active_support/concern"
 
 module RenderKit
   module RenderingHelperMonkeyPatch
-    def render(options = {}, locals = {}, &)
+    def render(options = {}, locals = {}, &block)
       if renderable = RenderKit::RenderableRegistry.get_renderables(options)
-        renderable.new(locals).render_in(self, &)
+        renderable.new(**locals).render_in(self, &block)
       else
         super
       end