+++ test/sandbox/app/components/test_implicit_controller_render_component.rb
@@ -2,4 +2,8 @@
 
 class TestImplicitControllerRenderComponent < ViewComponent::Base
   ActionView::Base.register_renderable("test_implicit_controller_render", TestImplicitControllerRenderComponent, controller_render=true)
+
+  def initialize(title: nil)
+    @title = title
+  end
 end