+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Update Documentation
+- Add support for params in controller and locals in view
 
 ## [0.1.0] - 2024-08-20